+++ src/lib.rs
@@ -37,7 +37,7 @@
 //!
 //! - `build.rs`, build script that places `device.x` somewhere the linker can find.
 //! - `device.x`, linker script that weakly aliases all the interrupt handlers to the default
-//! exception handler (`DefaultHandler`).
+//!   exception handler (`DefaultHandler`).
 //! - `lib.rs`, the generated code.
 //!
 //! All these files must be included in the same device crate. The `lib.rs` file contains several
@@ -95,7 +95,7 @@
 //!
 //! - `build.rs`, build script that places `device.x` somewhere the linker can find.
 //! - `device.x`, linker script that weakly aliases all the interrupt handlers to the default
-//! exception handler (`DefaultHandler`).
+//!   exception handler (`DefaultHandler`).
 //! - `lib.rs`, the generated code.
 //!
 //! All these files must be included in the same device crate. The `lib.rs` file contains several
@@ -149,22 +149,26 @@
 //! The resulting crate must provide an opt-in `rt` feature and depend on these crates:
 //!
 //! - [`critical-section`](https://crates.io/crates/critical-section) v1.x
-//! - [`riscv`](https://crates.io/crates/riscv) v0.9.x (if target is RISC-V)
-//! - [`riscv-rt`](https://crates.io/crates/riscv-rt) v0.9.x (if target is RISC-V)
+//! - [`riscv`](https://crates.io/crates/riscv) v0.12.x (if target is RISC-V)
+//! - [`riscv-peripheral`](https://crates.io/crates/riscv-peripheral) v0.2.x (if target is RISC-V and has standard peripherals)
+//! - [`riscv-rt`](https://crates.io/crates/riscv-rt) v0.13.x (if target is RISC-V)
 //! - [`vcell`](https://crates.io/crates/vcell) v0.1.x
 //!
-//! The `*-rt` dependencies must be optional only enabled when the `rt` feature is enabled. The
-//! `Cargo.toml` of the device crate will look like this for a RISC-V target:
+//! The `*-rt` dependencies must be optional only enabled when the `rt` feature is enabled.
+//! If target is RISC-V and supports vectored mode, you must include a feature `v-trap` to activate `riscv-rt/v-trap`.
+//! The `Cargo.toml` of the device crate will look like this for a RISC-V target:
 //!
 //! ``` toml
 //! [dependencies]
 //! critical-section = { version = "1.0", optional = true }
-//! riscv = "0.9.0"
-//! riscv-rt = { version = "0.9.0", optional = true }
+//! riscv = "0.12.1"
+//! riscv-peripheral = "0.2.0"
+//! riscv-rt = { version = "0.13.0", optional = true }
 //! vcell = "0.1.0"
 //!
 //! [features]
 //! rt = ["riscv-rt"]
+//! v-trap = ["rt", "riscv-rt/v-trap"]
 //! ```
 //!
 //! # Peripheral API
@@ -593,7 +597,7 @@ pub mod config;
 pub mod generate;
 pub mod util;
 
-pub use config::{Config, Target};
+pub use config::{Config, Settings, Target};
 
 #[non_exhaustive]
 pub struct Generation {